refactor(components):partition each component with it style instead
