+++ client/src/components/Home/InputBox/index.jsx
@@ -1,6 +1,7 @@
 import {useState} from 'react';
 import PropTypes from 'prop-types';
 
+import './index.css'
 
 const Input = ({ addPost }) => {
     const [input, setInput] = useState('');
