+++ client/src/components/Home/PostBox/index.css
@@ -1,0 +1,16 @@
+.Post div{
+    padding: 5px;
+    background-color: white;
+}
+
+.Title {
+    word-break: break-all;  
+}
+
+.Delete-post {
+    background: none;
+    border: none;
+    padding: 0;
+    float: right;
+    cursor: pointer;
+}
