+++ client/src/components/Movies/MovieBox/index.css
