+++ client/src/components/Navbar/index.jsx
@@ -1,4 +1,4 @@
-import '../styles/Navbar.css'
+import './index.css'
 
 const Navbar = () => {
     return (<div className="Navbar">
