+++ client/src/components/Review/index.jsx
@@ -1,0 +1,18 @@
+import Navbar from '../Navbar';
+import ReviewBox from './ReviewBox'
+import Footer from '../Footer';
+
+const Review = () =>  {
+  
+    return (
+        <div>
+        <Navbar></Navbar>
+        <ReviewBox >
+        </ReviewBox>
+        <Footer></Footer>
+          </div>
+    );
+  }
+  
+
+export default Review
