+++ client/src/main.css
@@ -1,0 +1,8 @@
+* {
+    box-sizing: border-box;
+    background-color: #dddddd;
+}
+
+body {
+    margin:0;
+}
