--- client/src/pages/Post.jsx
@@ -1,18 +1,0 @@
-import Navbar from '../components/Navbar';
-import Review from '../components/Review'
-import Footer from '../components/Footer';
-
-const Post = () =>  {
-  
-    return (
-        <div>
-        <Navbar></Navbar>
-        <Review >
-        </Review>
-        <Footer></Footer>
-          </div>
-    );
-  }
-  
-
-export default Post
